requirements.txt
fastapi==0.100.0
uvicorn==0.23.2
sqlalchemy==2.0.21
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
pydantic==2.3.0
pydantic-settings==2.0.3
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.11.3
psycopg2-binary==2.9.7
langchain==0.0.285
langchain-community==0.0.13
openai==0.27.8
pandas==2.0.3
numpy==1.24.3
faiss-cpu==1.7.4
slack-sdk==3.21.3
twilio==8.5.0
python-magic==0.4.27
aiofiles==23.2.1
openpyxl==3.1.2
pypdf2==3.0.1
docx2txt==0.8
unstructured==0.10.25
sentence-transformers==2.2.2